changed names and fixing build

+++ zhuravlev.ivan/I4/fill_array.cpp
@@ -1,11 +1,12 @@
-#include "fill.array.h"
+#include "fill_array.h"
+#include <cstddef>
 #include <iostream>
 
 void fillDynArray(int** array_for_filling, size_t nRows, size_t nColomns, std::fstream& input)
 {
-  for (int i = 0; i < nRows; ++i)
+  for (size_t i = 0; i < nRows; ++i)
   {
-    for (int j = 0; j < nColomns; ++j)
+    for (size_t j = 0; j < nColomns; ++j)
     {
       input >> array_for_filling[i][j];
       if (!input)
@@ -18,9 +19,9 @@ void fillDynArray(int** array_for_filling, size_t nRows, size_t nColomns, std::f
 
 void fillIntegratedArray(int *arr, size_t nRow, size_t nColomn, std::ifstream& input)
 {
-  for (int i = 0; i < nRow; ++i)
+  for (size_t i = 0; i < nRow; ++i)
   {
-    for (int j = 0; j < nColomn; ++j)
+    for (size_t j = 0; j < nColomn; ++j)
     {
       input >> arr[nRow * i + j];
       if (!input)
